:recycle: refactor: 상품 구매 요청 보기 url 변경 (#89)
